Modificación en presentación y dto_negocios
- Añadir clases HerramientaDTO y MaquinariaDTO, y modificar HerramientaIngresadaDTO.
- Modificar CoordinadorNegocio para instanciar las herramientas y maquinaria precargadas.
- Modificar forms de Herramienta y Maquinaria para mostrar en las tablas los nombres de herramienta y maquinaria.
- Enviar la lista de actividades a la fachada de actividades.

+++ dto_negocios/src/main/java/dto/BitacoraDTO.java
@@ -12,12 +12,12 @@ import java.time.LocalDate;
  */
 public class BitacoraDTO {
     LocalDate fecha;
-    Integer idObra;
+    Long idObra;
 
     public BitacoraDTO() {
     }
     
-    public BitacoraDTO(LocalDate fecha, Integer idObra) {
+    public BitacoraDTO(LocalDate fecha, Long idObra) {
         this.fecha = fecha;
         this.idObra = idObra;
     }
@@ -30,11 +30,11 @@ public class BitacoraDTO {
         this.fecha = fecha;
     }
 
-    public Integer getIdObra() {
+    public Long getIdObra() {
         return idObra;
     }
 
-    public void setIdObra(Integer idObra) {
+    public void setIdObra(Long idObra) {
         this.idObra = idObra;
     }
 }
